refactor: rework e2e tests
