+++ package.json
@@ -51,8 +51,7 @@
     "prettier": "^2.8.8",
     "ts-jest": "29.4.6",
     "ts-node": "10.9.2",
-    "typescript": "5.9.3",
-    "verdaccio": "^6.2.4"
+    "typescript": "5.9.3"
   },
   "nx": {
     "includedScripts": []
