+++ packages/nx-go-e2e/project.json
@@ -7,9 +7,7 @@
   "targets": {
     "e2e": {
       "options": {
-        "runInBand": true,
-        "silent": false,
-        "verbose": true
+        "runInBand": true
       },
       "dependsOn": ["^build"],
       "configurations": {
