--- packages/nx-go-e2e/shared/update-nx-config.ts
@@ -1,13 +1,0 @@
-import { updateFile } from '@nx/plugin/testing';
-import { join } from 'path';
-
-export const addNxTarget = (
-  directory: string,
-  target: string,
-  targetConfig: object
-): void =>
-  updateFile(join(directory, 'project.json'), (content) => {
-    const json = JSON.parse(content);
-    json['targets'][target] = targetConfig;
-    return JSON.stringify(json);
-  });
