+++ packages/nx-go-e2e/src/nx-go.spec.ts
@@ -1,46 +1,21 @@
 import {
   checkFilesExist,
+  cleanup,
+  ensureNxProject,
   readFile,
   readJson,
   runNxCommandAsync,
   uniq,
   updateFile,
 } from '@nx/plugin/testing';
-import { execSync } from 'child_process';
-import { rmSync } from 'fs';
 import { join } from 'path';
-import createTestProject from '../shared/create-test-project';
-import { addNxTarget } from '../shared/update-nx-config';
 
 describe('nx-go', () => {
   const appName = uniq('app');
   const libName = uniq('lib');
 
-  let projectDirectory: string;
-
-  beforeAll(() => {
-    projectDirectory = createTestProject();
-
-    // The plugin has been built and published to a local registry in the jest globalSetup
-    // Install the plugin built with the latest source code into the test repo
-    execSync(`npm install -D @nx-go/nx-go@latest`, {
-      cwd: projectDirectory,
-      stdio: 'inherit',
-      env: process.env,
-    });
-  });
-
-  afterAll(() => {
-    rmSync(projectDirectory, { recursive: true, force: true });
-  });
-
-  it('should be installed', () => {
-    // npm ls will fail if the package is not installed properly
-    execSync('npm ls @nx-go/nx-go', {
-      cwd: projectDirectory,
-      stdio: 'inherit',
-    });
-  });
+  beforeAll(() => ensureNxProject('@nx-go/nx-go', 'dist/packages/nx-go'));
+  afterAll(() => cleanup());
 
   it('should initialize the workspace', async () => {
     await runNxCommandAsync(`generate @nx-go/nx-go:init`);
@@ -79,11 +54,8 @@ describe('nx-go', () => {
 
   it('should create an application in a sub directory', async () => {
     const name = uniq('app');
-    // directory is not derived since Nx 20
-    const nxVersion = parseInt(process.env.NX_VERSION) >= 20;
-    const directory = nxVersion ? 'apps' : `apps/${name}`;
     await runNxCommandAsync(
-      `g @nx-go/nx-go:application --name=${name} --directory=${directory}`
+      `g @nx-go/nx-go:application --name=${name} --directory=apps/${name}`
     );
 
     expect(() => checkFilesExist(`apps/${name}/main.go`)).not.toThrow();
@@ -120,9 +92,13 @@ describe('nx-go', () => {
   });
 
   describe('Generate', () => {
-    beforeAll(() =>
-      addNxTarget(appName, 'generate', { executor: '@nx-go/nx-go:generate' })
-    );
+    beforeAll(() => {
+      updateFile(join(appName, 'project.json'), (content) => {
+        const json = JSON.parse(content);
+        json['targets']['generate'] = { executor: '@nx-go/nx-go:generate' };
+        return JSON.stringify(json);
+      });
+    });
 
     it('should execute go generate', async () => {
       const result = await runNxCommandAsync(`run ${appName}:generate`);
