+++ packages/nx-go/src/graph/create-nodes-v2.ts
