+++ packages/nx-go/src/index.ts
@@ -1,6 +1,6 @@
 import { NxPlugin } from '@nx/devkit';
 import { createDependencies } from './graph/create-dependencies';
-import { createNodesV2 } from './graph/create-nodes';
+import { createNodesV2 } from './graph/create-nodes-v2';
 
 const plugin: NxPlugin = {
   name: '@nx-go/nx-go',
