+++ project.json
@@ -1,14 +1,4 @@
 {
   "name": "@nx-go/source",
-  "$schema": "node_modules/nx/schemas/project-schema.json",
-  "targets": {
-    "local-registry": {
-      "executor": "@nx/js:verdaccio",
-      "options": {
-        "port": 4873,
-        "config": ".verdaccio/config.yml",
-        "storage": "tmp/local-registry/storage"
-      }
-    }
-  }
+  "$schema": "node_modules/nx/schemas/project-schema.json"
 }
